server: Cleanup HTTP routes
Change-type: patch
