+++ lib/server/http/middlewares.js
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+const _ = require('lodash')
 const randomstring = require('randomstring')
 const bodyParser = require('body-parser')
 const compression = require('compression')
@@ -24,7 +25,7 @@ const logger = require('../../logger').getLogger(__filename)
 // A regex that matches file types that should be compressed
 const COMPRESSION_REGEX = /\.(mp3|js|map|svg)$/
 
-module.exports = (application) => {
+module.exports = (application, jellyfish) => {
 	// Enable compression only for static files and the homepage
 	application.use(compression({
 		filter: (request, response) => {
@@ -86,4 +87,11 @@ module.exports = (application) => {
 			time
 		})
 	}))
+
+	application.use((request, response, next) => {
+		const authorization = request.headers.authorization
+		const token = _.last(_.split(authorization, ' '))
+		request.sessionToken = token || jellyfish.sessions.guest
+		return next()
+	})
 }
