+++ lib/server/routes.js
@@ -14,381 +14,10 @@
  * limitations under the License.
  */
 
-const Bluebird = require('bluebird')
 const randomstring = require('randomstring')
-const errio = require('errio')
-const sync = require('../sync')
-const environment = require('../environment')
-const logger = require('../logger').getLogger(__filename)
-const fs = require('fs')
 const _ = require('lodash')
-const multer = require('multer')
-const path = require('path')
-const uuid = require('uuid/v4')
-const {
-	version,
-	codename
-} = require('../../package')
-const Storage = require('./file-storage')
-const errorReporter = require('./error-reporter')
 
-const fileStore = new Storage({
-	driver: process.env.FS_DRIVER || 'localFS'
-})
-const upload = multer({
-	storage: multer.memoryStorage()
-})
-const rawChangelog = fs.readFileSync('./CHANGELOG.md', 'utf8')
-
-// Responds with a 400 error and a message
-const error400 = (message, response) => {
-	return response.status(400).json({
-		error: true,
-		data: message
-	})
-}
-
-// Generic wrapper for responding with data from endpoint handlers
-const respond = (handlerPromise, response) => {
-	return handlerPromise
-		.then((data) => {
-			response.status(200).json({
-				error: false,
-				data
-			})
-		}).catch((error) => {
-			response.status(500).json({
-				error: true,
-				data: error.message
-			})
-		})
-}
-
-// Gets a query schema from a schema, view card or view id
-const loadSchema = (jellyfish, input, token, ctx) => {
-	return Bluebird.try(() => {
-		if (_.isString(input)) {
-			// Try and load the view by id first
-			return jellyfish.getCardById(ctx, token, input, {
-				type: 'view'
-			})
-				.then((viewCardFromId) => {
-					if (!viewCardFromId || viewCardFromId.type !== 'view') {
-						// Now try and load the view by slug
-						return jellyfish.getCardBySlug(ctx, token, input, {
-							type: 'view'
-						})
-							.then((viewCardFromSlug) => {
-								if (!viewCardFromSlug || viewCardFromSlug.type !== 'view') {
-									throw new jellyfish.errors.JellyfishNoView(`Unknown view: ${input}`)
-								}
-								return viewCardFromSlug
-							})
-					}
-					return viewCardFromId
-				})
-		}
-
-		return input
-	})
-}
-
-// Strip the semantic version header from the changelog, and only use the last
-// 10 version bumps
-const changelog = rawChangelog.slice(rawChangelog.indexOf('##'))
-	.split('##')
-	.slice(0, 10)
-	.join('##')
-
-exports.bindRoutes = (jellyfish, queue, app, socketServer) => {
-	// A middleware that retrieves a Bearer token and attach it to the request object
-	// as `sessionToken`
-	const authorize = (request, response, next) => {
-		const authorization = request.headers.authorization
-
-		const token = _.last(_.split(authorization, ' '))
-
-		request.sessionToken = token || jellyfish.sessions.guest
-
-		return next()
-	}
-
-	app.get('/', (request, response) => {
-		response.sendFile(path.join('dist', 'index.html'))
-	})
-
-	app.get('/api/v2/config', (request, response) => {
-		response.send({
-			changelog,
-			codename,
-			version
-		})
-	})
-
-	/*
-	 * Queries the database for cards and returns them as an array.
-	 * Accepts either a serialized JSON schema as qs params, or a view id or view card as the `query` parameter.
-	 */
-	app.post('/api/v2/query', authorize, (request, response) => {
-		const ctx = request.context
-
-		if (_.isEmpty(request.body)) {
-			return error400('No query schema', response)
-		}
-
-		const {
-			query,
-			options
-		} = request.body
-		options.ctx = ctx
-
-		const handler = loadSchema(jellyfish, query, request.sessionToken, ctx)
-			.then((schema) => {
-				return jellyfish.query(ctx, request.sessionToken, schema, options)
-			})
-
-		return respond(handler, response)
-	})
-
-	// Some services, such as Workable, require the user to register
-	// different endpoints for every type of event we're interested in,
-	// which means we can't send more than one event type to
-	// /api/v2/hooks/workable. As a solution, we can allow this rule to
-	// have an optional "type" parameter that is not used for anything
-	// apart from differentiating the endpoints.
-	app.all('/api/v2/hooks/:provider/:type*?', (request, response) => {
-		const hostname = request.headers.host
-		logger.info(request.context, 'Received webhook', {
-			source: request.params.provider
-		})
-
-		const integrationToken = environment.getIntegrationToken(
-			request.params.provider)
-
-		if (!sync.isValidExternalEventRequest(
-			integrationToken, request.params.provider, request.rawBody, request.headers)) {
-			logger.warn(request.context, 'Webhook rejected', {
-				source: request.params.provider,
-				hostname,
-				body: request.body
-			})
-
-			return response.status(401).json({
-				error: true,
-				data: 'Webhook rejected'
-			})
-		}
-
-		if (_.isEmpty(request.body)) {
-			return error400('Invalid external event', response)
-		}
-		const ctx = request.context
-
-		const EXTERNAL_EVENT_TYPE = 'external-event'
-		const suffix = randomstring.generate().toLowerCase()
-
-		return jellyfish.getCardBySlug(ctx, jellyfish.sessions.admin, EXTERNAL_EVENT_TYPE, {
-			type: 'type'
-		}).then((typeCard) => {
-			if (!typeCard) {
-				throw new Error(`No type card: ${EXTERNAL_EVENT_TYPE}`)
-			}
-
-			return queue.enqueue(jellyfish.sessions.admin, {
-				action: 'action-create-card',
-				card: typeCard.id,
-				type: typeCard.type,
-				context: ctx,
-				arguments: {
-					properties: {
-						slug: `${EXTERNAL_EVENT_TYPE}-${suffix}`,
-						version: '1.0.0',
-						data: {
-							source: request.params.provider,
-							headers: request.headers,
-							payload: request.body
-						}
-					}
-				}
-			})
-		}).then((requestObject) => {
-			return response.status(200).json({
-				error: false,
-				data: requestObject
-			})
-		}).catch((error) => {
-			return response.status(500).json({
-				error: true,
-				data: {
-					type: 'Error',
-					message: error.message
-				}
-			})
-		})
-	})
-
-	app.get('/api/v2/file/:cardId/:fileName', authorize, (request, response) => {
-		const {
-			cardId,
-			fileName
-		} = request.params
-
-		const card = jellyfish.getCardById(request.context, request.sessionToken, cardId)
-
-		if (!card) {
-			response.send(404)
-		}
-
-		fileStore.retrieve(cardId, fileName).then((file) => {
-			return response.status(200).send(file)
-		}).catch((error) => {
-			return response.status(500).json({
-				error: true,
-				data: {
-					type: 'Error',
-					message: error.message
-				}
-			})
-		})
-	})
-
-	app.post('/api/v2/action', authorize, upload.any(), async (request, response) => {
-		const ctx = request.context
-
-		// If files are uploaded, the action payload is serialized as the form field
-		// "action" and will need to be parsed
-		const action = request.files ? JSON.parse(request.body.action) : request.body
-
-		logger.info(ctx, 'HTTP action request', {
-			action
-		})
-
-		if (_.isEmpty(action)) {
-			return error400('No action request', response)
-		}
-
-		const files = []
-
-		if (request.files) {
-			// Upload magic
-			request.files.forEach((file) => {
-				const name = `${uuid()}.${file.originalname}`
-				_.set(action.arguments.properties, file.fieldname, name)
-				files.push({
-					buffer: file.buffer,
-					name
-				})
-			})
-		}
-
-		action.context = ctx
-		return queue.enqueue(request.sessionToken, action).then((id) => {
-			return queue.waitResults(ctx, id)
-		}).then(async (results) => {
-			logger.info(ctx, 'Got action results', results)
-			if (!results.error && request.files) {
-				const cardId = results.data.id
-
-				for (const item of files) {
-					await fileStore.store(cardId, item.name, item.buffer)
-				}
-			}
-			return results
-		}).then((results) => {
-			if (results.error) {
-				errorReporter.reportException(ctx, errio.fromObject(results.data))
-			}
-
-			const code = results.error ? 500 : 200
-			return response.status(code).json(results)
-		}).catch((error) => {
-			const currentDate = new Date()
-			const errorObject = errio.toObject(error, {
-				stack: true
-			})
-
-			logger.error(ctx, 'HTTP unexpected error', {
-				error: errorObject
-			})
-
-			errorReporter.reportException(ctx, error)
-			return response.status(500).json({
-				error: true,
-				timestamp: currentDate.toISOString(),
-				data: errorObject
-			})
-		})
-	})
-
-	app.get('/status', (request, response) => {
-		const currentDate = new Date()
-
-		return Bluebird.props({
-			queue: queue.getStatus(),
-			kernel: jellyfish.getStatus()
-		}).then((status) => {
-			return response.status(200).json(status)
-		}).catch((error) => {
-			const errorObject = errio.toObject(error, {
-				stack: true
-			})
-
-			logger.error(request.context, 'Status error', {
-				error: errorObject
-			})
-
-			errorReporter.reportException(request.context, error)
-			return response.status(500).json({
-				error: true,
-				timestamp: currentDate.toISOString(),
-				data: errorObject
-			})
-		})
-	})
-
-	/*
-	* Health-Check endpoint
-	*/
-	app.get('/ping', (request, response) => {
-		const currentDate = new Date()
-		const session = jellyfish.sessions.admin
-
-		return queue.enqueue(session, {
-			action: 'action-create-card',
-			card: 'ping',
-			type: 'type',
-			currentDate,
-			context: request.context,
-			arguments: {
-				properties: {
-					data: {
-						timestamp: currentDate.toISOString()
-					}
-				}
-			}
-		}).then((id) => {
-			return queue.waitResults(request.context, id)
-		}).then(async (results) => {
-			const code = results.error ? 500 : 200
-			return response.status(code).json(results)
-		}).catch((error) => {
-			const errorObject = errio.toObject(error, {
-				stack: true
-			})
-
-			logger.error(request.context, 'Ping error', {
-				error: errorObject
-			})
-
-			errorReporter.reportException(request.context, error)
-			return response.status(500).json({
-				error: true,
-				timestamp: currentDate.toISOString(),
-				data: errorObject
-			})
-		})
-	})
-
+exports.bindRoutes = (jellyfish, app, socketServer) => {
 	const openStreams = {}
 
 	app.once('close', () => {
