+++ lib/server/web-server.js
@@ -23,6 +23,7 @@ const http = require('http')
 const socketIo = require('socket.io')
 
 const middlewares = require('./http/middlewares')
+const routes = require('./http/routes')
 const {
 	bindRoutes
 } = require('./routes')
@@ -33,8 +34,11 @@ exports.createWebServer = (context, jellyfish, queue, configuration) => {
 	const socketServer = socketIo(server)
 
 	application.set('port', configuration.http.port)
-	middlewares(application)
-	bindRoutes(jellyfish, queue, application, socketServer)
+	middlewares(application, jellyfish)
+	routes(application, jellyfish, queue)
+
+	// Socket server routes
+	bindRoutes(jellyfish, application, socketServer)
 
 	// Resolve the promise once the server is ready to accept incoming
 	// connections.
